SideKicks

Documentation for SideKicks, a Julia package for analyzing the observed properties of stellar binaries containing a compact object, in order to perform parameter inference on the supernova mass loss and natal kick.

Package features

# TODO - RTW

Index

SideKicks.SideKicks

SideKicks.KickMCMCResults

SideKicks.ModVonMises

SideKicks.Observations

SideKicks.Priors

SideKicks.WrappedCauchy

SideKicks.RV_semiamplitude_K1

SideKicks.arbitraryEjectaBH

SideKicks.createGeneralMCMCModel

SideKicks.createSimplifiedMCMCModel

SideKicks.create_1D_density

SideKicks.create_2D_density

SideKicks.create_compound_1D_densities

SideKicks.get_bounds_for_fractions

SideKicks.kepler_P_from_a

SideKicks.kepler_a_from_P

SideKicks.post_supernova_circular_orbit_P

SideKicks.post_supernova_circular_orbit_a

SideKicks.post_supernova_circular_orbit_vsys

SideKicks.post_supernova_general_orbit_parameters

SideKicks.relative_velocity

SideKicks.restrictedEjectaBH

SideKicks.SideKicks — Module

Main module for SideKicks.jl – a analysis package for performing parameter inference on compact object stellar binaries.

SideKicks.KickMCMCResults — Type

mutable struct KickMCMCResults

KickMCMCResults contains the MCMC model and Observations structs, the Results dict, the chains that resulted from the MCMC, and the parameters that went into the sampler.

SideKicks.ModVonMises — Type

struct ModVonMises{T1<:Real, T2<:Real} <: ContinuousUnivariateDistribution

This is just a wrapper on top of the VonMises distribution (as defined in Distributions.jl) to extend its domain. This is because the domain of VonMises is defined to be [μ-π, μ+π], and the angles we are concerned with range from [0,2π]

SideKicks.Observations — Type

mutable struct Observations

Observations contains the symbols, values, errors, and units of each observed parameter.

SideKicks.Priors — Type

mutable struct Priors

Priors contains the prior distribution of each of the desired parameters

RTW: check that N and E are identical to RA and Dec...

SideKicks.WrappedCauchy — Type

struct WrappedCauchy{T1<:Real, T2<:Real} <: ContinuousUnivariateDistribution

The WrappedCauchy distribution resembles the Cauchy distribution defined on the unit circle from 0 to 2π, with the endpoints wrapped back to each other.

SideKicks.RV_semiamplitude_K1 — Method

RV_semiamplitude_K1(;m1, m2, P, e, i)

Compute the amplitude of radial velocity variations given orbital parameters and masses

Arguments:

m1: mass of observed star [g]

m2: mass of companion [g]

P: orbital period [s]

e: orbital eccentricity [-]

i: orbital inclination [rad]

Output:

K1: amplitude of radial velocity variation of star 1 [cm/s]

SideKicks.arbitraryEjectaBH — Method

arbitraryEjectaBH(m2_i, frac)

#TODO Description

Arguments:

#TODO

m2_i:

frac:

Output:

#TODO

SideKicks.createGeneralMCMCModel — Method

createGeneralMCMCModel(observations, observed_values, observed_errors)

Create a Turing model to perform an MCMC sampling of the pre-explosion and kick properties of a system, assuming pre-explosion eccentricity.

RTW does acos just work? Do I need to worry about domain/range issues?

Check velocities, the conversions are a bit funky

Arguments:

observations: the parameters taken from observations [Vector{Symbol}]

observed_values: the values of the parameters [Vector{Float64}]

observed_errors: the errors of the observations [Vector{Float64}]

Output:

kickmodel: A Turing model for sampling

SideKicks.createSimplifiedMCMCModel — Method

createSimplifiedMCMCModel(observations, observed_values, observed_errors)

Description Create a Turing model to perform an MCMC sampling of the pre-explosion and kick properties of a system, assuming pre-explosion circularity.

RTW this is more simplistic than just using a circular model, it's also assuming you know the eccentricity and don't care about radial velocity etc.

Arguments:

observations: the parameters taken from observations [Vector{Symbol}]

observed_values: the values of the parameters [Vector{Float64}]

observed_errors: the errors of the observations [Vector{Float64}]

Output:

kickmodel: A Turing model for sampling

SideKicks.create_1D_density — Method

create_1D_density(values, range, chain_weights, fraction_1D, axis, nbins)

#TODO Description

Arguments:

#TODO

values:

range:

chain_weights:

fraction_1D:

axis:

nbins:

Output:

#TODO

SideKicks.create_2D_density — Method

create_2D_density(values1, ranges1, values2, ranges2, chain_weights, fractions, axis, nbins)

#TODO Description

Arguments:

#TODO

values1:

ranges1:

values2:

ranges2:

weights:

fractions:

axis:

nbins:

Output:

#TODO

SideKicks.create_compound_1D_densities — Method

create_1D_density(values, range, chain_weights, fraction_1D, axis, nbins)

#TODO Description

Arguments:

#TODO

values:

range:

chain_weights:

fraction_1D:

axis:

nbins:

Output:

#TODO

SideKicks.get_bounds_for_fractions — Method

get_bounds_for_fractions(h, fractions)

#TODO Description

Arguments:

#TODO h: fractions:

Output:

#TODO

SideKicks.kepler_P_from_a — Method

kepler_P_from_a(;m1, m2, a)

Obtain period from semimajor axis using Kepler's third law

Arguments:

m1: mass of first companion [g]

m2: mass of 2nd companion [g]

a: semi-major axis of the orbit [cm]

Output:

P: the orbital period [s]

SideKicks.kepler_a_from_P — Method

kepler_a_from_P(;m1, m2, P)

Obtain semimajor axis from period using Kepler's third law

Arguments:

m1: mass of first companion [g]

m2: mass of 2nd companion [g]

P: orbital period [s]

Output:

a: semi-major axis of the orbit [cm]

SideKicks.post_supernova_circular_orbit_P — Method

post_supernova_circular_orbit_P(;m1, m2, P, m1_f=-1, m2_f, vkick=0, θ=0, ϕ=0, vimp=0)

Same as post_supernova_circular_orbit_a, except that it receives the initial orbital period as input and returns the final orbital period and eccentricity.

Arguments:

m1: pre-explosion mass of non-exploding component [g]

m2: pre-explosion mass of exploding component [g]

P: pre-explosion orbital period [d]

m1_f: post-explosion mass of non-exploding component [g]

m2_f: post-explosion mass of exploding component [g]

vkick: kick velocity [cm/s]

θ: polar kick angle (away from e_par) [rad]

ϕ: azimuthal kick angle (off of e_perp) [rad]

vimp: imparted kick velocity on companion [cm/s]

Output:

P_f: post-explosion orbital period [d]

e_f: post-explosion excentricity [-]

SideKicks.post_supernova_circular_orbit_a — Method

post_supernova_circular_orbit_a(;m1, m2, a, m1_f=-1, m2_f, vkick=0, θ=0, ϕ=0, vimp=0)

Compute post-kick properties for a circular pre-explosion orbit. Equivalent to Tauris et al. (1999): Monthly Notices of the Royal Astronomical Society, Volume 310, Issue 4, pp. 1165-1169.

Arguments:

m1: pre-explosion mass of non-exploding component [g]

m2: pre-explosion mass of exploding component [g]

a: pre-explosion orbital separation [cm]

m1_f: post-explosion mass of non-exploding component [g]

m2_f: post-explosion mass of exploding component [g]

vkick: kick velocity [cm/s]

θ: polar kick angle (away from e_par) [rad]

ϕ: azimuthal kick angle (off of e_perp) [rad]

vimp: imparted kick velocity on companion [cm/s]

Output:

a_f: post-explosion orbital separation [cm]

e_f: post-explosion excentricity [-]

SideKicks.post_supernova_circular_orbit_vsys — Method

post_supernova_circular_orbit_vsys(;m1, m2, a, m1_f=-1, m2_f, vkick=0, θ=0, ϕ=0, vimp=0)

Compute post-kick properties for a circular pre-explosion orbit using equations from Tauris & Takens (1999)

Arguments:

m1: pre-explosion mass of non-exploding component [g]

m2: pre-explosion mass of exploding component [g]

a: pre-explosion orbital separation [cm]

m1_f: post-explosion mass of non-exploding component [g]

m2_f: post-explosion mass of exploding component [g]

vkick: kick velocity [cm/s]

θ: polar kick angle (away from e_par) [rad]

ϕ: azimuthal kick angle (off of e_perp) [rad]

vimp: imparted kick velocity on companion [cm/s]

Output:

vsys_f: post-explosion systemic velocity [cm/s]

SideKicks.post_supernova_general_orbit_parameters — Method

post_supernova_general_orbit_parameters(;m1, m2, a, e=0, m1_f=-1, m2_f, vkick=0, θ=0, ϕ=0, vimp=0,
    ν=0, Ω=0, ω=0, i=0)

Compute post-kick properties for a general pre-explosion orbit using equations from [Marchant, Willcox, Vigna-Gomez] TODO

Arguments:

m1: pre-explosion mass of non-exploding component [g]

m2: pre-explosion mass of exploding component [g]

a: pre-explosion orbital separation [cm]

e: pre-explosion orbital eccentricity [-]

m1_f: post-explosion mass of non-exploding component [g]

m2_f: post-explosion mass of exploding component [g]

vkick: kick velocity [cm/s]

θ: polar kick angle (away from e_par) [rad]

ϕ: azimuthal kick angle (off of e_perp) [rad]

vimp: imparted kick velocity on companion [cm/s]

Initial orbital orientation angles:

ν: true anomaly [rad]

Ω: pre-explosion longitude of the ascending node [rad]

ω: pre-explosion argument of periastron [rad]

i: pre-explosion inclination [rad]

Output: RTW: check!

a_f: post-explosion orbital separation [cm]

e_f: post-explosion orbital eccentricity [-]

Ω_f: post-explosion longitude of ascending node [rad]

ω_f: post-explosion argument of periastron [rad]

i_f: post-explosion inclination [rad]

v_n: post-explosion systemic velocity, toward N [rad]

v_w: post-explosion systemic velocity, toward W [rad]

v_rad: post-explosion radial velocity, toward O [rad]

SideKicks.relative_velocity — Method

relative_velocity(;m1, m2, a)

Calculate the relative orbital velocity for a circular orbit.

Arguments:

m1: mass of first companion [g]

m2: mass of 2nd companion [g]

a: semi-major axis of the orbit [cm]

Output:

v_rel: the relative velocity [cm/s]

SideKicks.restrictedEjectaBH — Method

restrictedEjectaBH(m2_i, frac; Ma=10, Mb=15, max_frac=1.0, min_frac=0.1)

#TODO Description

Arguments:

#TODO

m2_i:

frac:

Ma=10:

Mb=15:

max_frac=1.0:

min_frac=0.1:

Output:

#TODO